+++ src/constants.ts
@@ -1,1 +1,3 @@
 export const EXTENSION_PREFIX = 'sorbetto';
+
+export const PACKAGE_FILENAME = '__package.rb';